Separate delta processing and useful helpers
These helpers can come handy while processing healing.
